Retsructured tests and made them all pass. Todo: Clean up imports; reduce repetition.
